requirements.txt
# Modules pour l'exécution du projet
colorama==0.4.6
matplotlib==3.8.1
numpy==1.26.2
art==6.1

# Modules pour l'exécution des tests
pyyaml==6.0.1
requests==2.31.0
pytest==6.2.5
mock==4.0.3
pytest-html==3.2.0
pytest-mock==3.12.0
pytest-datadir-ng==1.1.1
pytest-pythonpath==0.7.3
pytest-timeout==2.2.0
pytest-cov==4.1.0
pytest-metadata==2.0.4
pytest==6.2.5